requirements.txt
DateTime==4.3
folium==0.12.1
geopandas==0.9.0
numpy==1.21.1
pandas==1.3.1
plotly-express==0.4.1
streamlit==0.86.0
streamlit-folium==0.4.0

